Rename screens to match routing paths
